+++ infra/sandbox/sandbox_web/src/apis/files/index.ts
@@ -2,3 +2,4 @@
  * File API 导出
  */
 export * from './api';
+export * from './types';